refactor(telemetry): migrate telemetry package to internal/stats/opentelemetry
It moves the package from `pkg/util/telemetry` to
`internal/stats/opentelemetry`.
