+++ cmd/varlogadm/cli.go
@@ -13,9 +13,9 @@ import (
 	"github.com/kakao/varlog/internal/admin/snwatcher"
 	"github.com/kakao/varlog/internal/buildinfo"
 	"github.com/kakao/varlog/internal/flags"
+	"github.com/kakao/varlog/internal/stats/opentelemetry"
 	"github.com/kakao/varlog/pkg/types"
 	"github.com/kakao/varlog/pkg/util/log"
-	"github.com/kakao/varlog/pkg/util/telemetry"
 )
 
 func newAdminApp() *cli.App {
@@ -115,11 +115,11 @@ func start(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	mp, stop, err := telemetry.NewMeterProvider(meterProviderOpts...)
+	mp, stop, err := opentelemetry.NewMeterProvider(meterProviderOpts...)
 	if err != nil {
 		return err
 	}
-	telemetry.SetGlobalMeterProvider(mp)
+	opentelemetry.SetGlobalMeterProvider(mp)
 	defer func() {
 		ctx, cancel := context.WithTimeout(context.Background(), c.Duration(flags.TelemetryExporterStopTimeout.Name))
 		defer cancel()
