+++ cmd/varlogmr/metadata_repository.go
@@ -13,9 +13,9 @@ import (
 	"github.com/kakao/varlog/internal/buildinfo"
 	"github.com/kakao/varlog/internal/flags"
 	"github.com/kakao/varlog/internal/metarepos"
+	"github.com/kakao/varlog/internal/stats/opentelemetry"
 	"github.com/kakao/varlog/pkg/types"
 	"github.com/kakao/varlog/pkg/util/log"
-	"github.com/kakao/varlog/pkg/util/telemetry"
 )
 
 func main() {
@@ -54,11 +54,11 @@ func start(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	mp, stop, err := telemetry.NewMeterProvider(meterProviderOpts...)
+	mp, stop, err := opentelemetry.NewMeterProvider(meterProviderOpts...)
 	if err != nil {
 		return err
 	}
-	telemetry.SetGlobalMeterProvider(mp)
+	opentelemetry.SetGlobalMeterProvider(mp)
 	defer func() {
 		ctx, cancel := context.WithTimeout(context.Background(), c.Duration(flags.TelemetryExporterStopTimeout.Name))
 		defer cancel()
