+++ internal/stats/opentelemetry/config.go
@@ -1,4 +1,4 @@
-package telemetry
+package opentelemetry
 
 import (
 	"go.opentelemetry.io/contrib/instrumentation/runtime"
