+++ internal/stats/opentelemetry/exporter.go
@@ -1,4 +1,4 @@
-package telemetry
+package opentelemetry
 
 import (
 	"context"
