+++ internal/stats/opentelemetry/metric_provider.go
@@ -1,4 +1,4 @@
-package telemetry
+package opentelemetry
 
 import (
 	"context"
